Changed search handling to do two searches
* search for any tiddlers with the same title
* search for tiddlers from people who follow you

The searches are done in serial so the button appears and then has
its class modified if there are more results. The counts are
included as data attributes in case somebody wants to do something
with that.

The searches are done as text because those results are delivered more
quickly.