Remove need to specify status.js in HtmlJavascript
The only script required now is the twiddlerscount.js script.  This checks for the presence of the status object and loads the script if it is not there, then it initialises.

Most of the time on TiddlySpace, the status object is already present but it is loaded after twiddlerscount.js is.  Perhaps this script should be included in the server side templates to guarantee load order? cc @cdent