Only show button and do second search when there are some results.
We don't want the button to display if there are no meaningful results,
and if there are no results there's no reason to do the second search

Meaningful is defined as more than one, as the existing tiddler will be
in the result set.

I adjusted one existing test and added another to try to confirm the
search behavior.