Fix #4 so current tiddlers does not show up in list of twins
The reason it was is because as @pads correctly pointed out
the filter was receiving a this.bag of undefined. This is because
the filtering was being called before the current tiddler ever
finished loading.

I reordered things so that the search is not done and filtered
until after the tiddler is loaded, thus ensuring that `this.bag`
is set properly.